+++ crates/containerd-shim-wasm/src/shim/cli.rs
@@ -13,7 +13,7 @@
 //! The shim can be configured using the [`Config`] struct:
 //!
 //! ```rust, no_run
-//! use containerd_shim_wasm::Config;
+//! use containerd_shim_wasm::shim::Config;
 //!
 //! let config = Config {
 //!     // Disable automatic logger setup
@@ -27,20 +27,13 @@
 //! };
 //! ```
 //!
-//! ## Version Information
-//!
-//! The module provides two macros for version information:
-//!
-//! - [`version!()`](crate::version) - Returns the crate version from Cargo.toml and
-//!   Git revision hash, if available.
-//!
 //! ## Example usage:
 //!
 //! ```rust, no_run
 //! use containerd_shim_wasm::{
-//!     Cli, version,
-//!     shim::{Shim, Sandbox, RuntimeContext, Version},
-//!     Config,
+//!     shim::{Shim, Cli, Config, Version, version},
+//!     sandbox::Sandbox,
+//!     sandbox::context::RuntimeContext,
 //! };
 //! use anyhow::Result;
 //!
@@ -83,8 +76,7 @@
 //! - `OTEL_SDK_DISABLED`: Disable OpenTelemetry SDK
 //!
 
-use crate::Config;
-use crate::shim::{Instance, Shim};
+use crate::shim::{Config, Instance, Shim};
 
 mod private {
     pub trait Sealed {}
