+++ crates/containerd-shim-wasmer/src/main.rs
@@ -1,4 +1,4 @@
-use containerd_shim_wasm::Cli;
+use containerd_shim_wasm::shim::Cli;
 use containerd_shim_wasmer::WasmerShim;
 
 fn main() {
