+++ crates/containerd-shim-wasmtime/src/instance.rs
@@ -2,11 +2,11 @@ use std::hash::Hash;
 use std::sync::LazyLock;
 
 use anyhow::{Context, Result, bail};
-use containerd_shim_wasm::sandbox::WasmLayer;
-use containerd_shim_wasm::shim::{
-    Compiler, Entrypoint, RuntimeContext, Sandbox, Shim, Version, WasmBinaryType,
+use containerd_shim_wasm::sandbox::Sandbox;
+use containerd_shim_wasm::sandbox::context::{
+    Entrypoint, RuntimeContext, WasmBinaryType, WasmLayer,
 };
-use containerd_shim_wasm::version;
+use containerd_shim_wasm::shim::{Compiler, Shim, Version, version};
 use tokio_util::sync::CancellationToken;
 use wasi_preview1::WasiP1Ctx;
 use wasi_preview2::bindings::Command;
